requirements.txt
Flask==3.1.1
pycryptodome==3.22.0